@@ -129,23 +109,20 @@
     "lodash.sortby": "^4.7.0",
     "lodash.truncate": "^4.4.2",
     "lodash.xor": "^4.5.0",
-    "marked": "^0.3.5",
     "material-ui": "^0.20.0",
-    "@material-ui/core": "^3.2.2",
     "material-ui-persian-date-picker-utils": "^0.1.2",
+    "meedan-check-react-jsonschema-form-material-ui-v1": "^30.0.3",
     "node-fetch": "^1.7.1",
-    "node-window": "^1.0.1",
     "nyc": "^13.1.0",
     "polished": "^1.2.1",
     "prop-types": "^15.6.2",
     "raw-loader": "^0.5.1",
     "rc-tooltip": "^3.5.0",
-    "react": "^16.4.1",
-    "react-chatview": "^0.2.4",
+    "react": "^16.13.1",
     "react-color": "^2.17.0",
     "react-copy-to-clipboard": "^5.0.0",
     "react-document-title": "^2.0.2",
-    "react-dom": "^16.4.1",
+    "react-dom": "^16.13.1",
     "react-dropzone": "^3.5.3",
     "react-emoji-render": "^0.4.5",
     "react-favicon": "0.0.8",
@@ -156,41 +133,26 @@
     "react-infinite-scroller": "^1.0.1",
     "react-intercom": "^1.0.15",
     "react-intl": "^2.2.3",
-    "react-jsonschema-form-material-ui": "1.0.30",
     "react-leaflet": "^1.5.0",
     "react-linkify": "^0.2.2",
     "react-redux": "^5.0.1",
-    "react-relay": "^1.0.0",
-    "react-router": "^3.0.0",
+    "react-relay": "^1.7.0",
+    "react-router": "^3.2.6",
     "react-router-redux": "^4.0.5",
     "react-select": "^1.3.0",
-    "react-transform-catch-errors": "^1.0.0",
-    "react-transform-hmr": "^1.0.1",
-    "react-truncate": "^2.1.0",
-    "react-video-js-player": "^1.1.1",
-    "redbox-react": "^1.0.1",
     "redux": "^3.0.0",
-    "redux-devtools": "^3.0.0-beta-3",
-    "redux-devtools-dock-monitor": "^1.0.0-beta-3",
-    "redux-devtools-log-monitor": "^1.0.0-beta-3",
-    "redux-form": "^6.3.2",
-    "redux-immutable-state-invariant": "^1.1.0",
-    "redux-logger": "^2.0.4",
     "redux-notify": "^0.2.0",
-    "redux-persist": "^4.0.0-beta1",
     "redux-thunk": "^2.1.0",
+    "relay-compiler": "^1.7.0",
     "rtl-detect": "^1.0.0",
     "serve-static": "^1.10.2",
-    "sinon": "^1.17.7",
-    "styled-components": "^2.1.0",
     "style-loader": "^0.13.1",
+    "styled-components": "^2.1.0",
     "superagent": "^3.3.0",
     "sync-request": "^4.0.3",
     "url-regex": "^3.2.0",
-    "watch": "^1.0.1",
-    "webpack": "^2.1.0-beta.25",
-    "webpack-bundle-analyzer": "^3.3.2",
-    "webpack-node-externals": "^1.5.4",
+    "video.js": "^7.7.5",
+    "webpack": "^2.7.0",
     "whatwg-fetch": "^2.0.1",
     "xregexp": "^3.1.1"
   }
