@@ -1,17 +1,13 @@
-import fs from 'fs';
-import request from 'sync-request';
-import gulp from 'gulp';
-import gutil from 'gulp-util';
-import rename from 'gulp-rename';
-import babel from 'gulp-babel';
-import concat from 'gulp-concat';
-import transifex from 'gulp-transifex';
-import jsonEditor from 'gulp-json-editor';
-import webpack from 'webpack';
-import mergeTransifex from './webpack/gulp-merge-transifex-translations';
-import webpackConfig from './webpack/config';
-import webpackServerConfig from './webpack/config_server';
-import buildConfig from './config-build';
+const fs = require('fs');
+const request = require('sync-request');
+const gulp = require('gulp');
+const gutil = require('gulp-util');
+const transifex = require('gulp-transifex');
+const jsonEditor = require('gulp-json-editor');
+const webpack = require('webpack');
+const mergeTransifex = require('./webpack/gulp-merge-transifex-translations');
+const webpackConfig = require('./webpack/config');
+const buildConfig = require('./config-build');
 
 let transifexClient = null;
 if (buildConfig.transifex) {
@@ -25,17 +21,6 @@ if (buildConfig.transifex) {
   });
 }
 
-gulp.task('replace-webpack-code', (callback) => {
-  [{
-    from: './webpack/replace/JsonpMainTemplate.runtime.js',
-    to: './node_modules/webpack/lib/JsonpMainTemplate.runtime.js',
-  }, {
-    from: './webpack/replace/log-apply-result.js',
-    to: './node_modules/webpack/hot/log-apply-result.js',
-  }].forEach(task => fs.writeFileSync(task.to, fs.readFileSync(task.from)));
-  callback();
-});
-
 gulp.task('relay:copy', (callback) => {
   if (buildConfig.relay.startsWith('http')) {
     const res = request('GET', buildConfig.relay);
@@ -47,19 +32,8 @@ gulp.task('relay:copy', (callback) => {
   }
 });
 
-gulp.task('webpack:build:server', (callback) => {
-  webpack(Object.create(webpackServerConfig), (err, stats) => {
-    if (err) {
-      gutil.log(err.message);
-      process.exit(1);
-    }
-    gutil.log('[webpack:build:server]', stats.toString({ colors: true, chunks: false }));
-    callback();
-  });
-});
-
 gulp.task('webpack:build:web', (callback) => {
-  webpack(Object.create(webpackConfig), (err, stats) => {
+  webpack(webpackConfig, (err, stats) => {
     if (err) {
       gutil.log(err.message);
       process.exit(1);
@@ -77,12 +51,7 @@ function copy_build_web_config_js() {
   return gulp.src('./config.js').pipe(gulp.dest('./build/web/js'));
 }
 
-function copy_build_web_config_test_js() {
-  return gulp.src('./test/config.js').pipe(gulp.dest('./build/web/js'));
-}
-
 gulp.task('copy:build:web', gulp.series(copy_build_web_assets, copy_build_web_config_js));
-gulp.task('copy:build:web:test', gulp.series(copy_build_web_assets, copy_build_web_config_test_js));
 
 gulp.task('transifex:download', () => {
   return gulp.src('./localization/transifex/**/*.json').pipe(transifexClient.pullResource());
@@ -113,43 +82,35 @@ gulp.task('transifex:languages', () => {
   return gulp.series();
 });
 
-gulp.task('build:web', gulp.series('replace-webpack-code', 'relay:copy', 'webpack:build:web', 'copy:build:web'));
-gulp.task('build:server', gulp.series('webpack:build:server'));
+gulp.task('build:web', gulp.series('relay:copy', 'webpack:build:web', 'copy:build:web'));
 
 // Dev mode — with 'watch' enabled for faster builds
 // Webpack only — without the rest of the web build steps.
-//
-const devConfig = Object.create(webpackConfig);
-
 gulp.task('webpack:build:web:dev', (callback) => {
-  // Enable watcher to monitor for changes
-  devConfig.watch = true;
+  const devConfig = {
+    ...webpackConfig,
+    bail: false, // don't stop on error
+    watch: true,
+  }
 
-  // Don't stop on error
-  devConfig.bail = false;
-
-  // Disable sourcemaps, for faster compile
-  // (Enable if needed, by commenting this out)
-  devConfig.devtool = 'eval';
-
-  webpack(Object.create(devConfig), (err, stats) => {
+  webpack(devConfig, (err, stats) => {
     if (err) {
-      throw new gutil.PluginError('webpack:build', err);
+      return callback(new gutil.PluginError('webpack:build', err));
     }
     gutil.log('[webpack:build:web:dev]', stats.toString({
       colors: true,
       hash: false,
       version: false,
       timings: true,
-      assets: false,
+      assets: true,
       chunks: false,
       modules: false,
       reasons: false,
       children: false,
       source: false,
       errors: true,
-      errorDetails: false,
-      warnings: false,
+      errorDetails: true,
+      warnings: true,
       publicPath: false,
     }));
   });
@@ -157,7 +118,7 @@ gulp.task('webpack:build:web:dev', (callback) => {
   // never call callback()
 });
 
-gulp.task('build:web:dev', gulp.series('replace-webpack-code', 'relay:copy', 'copy:build:web', 'webpack:build:web:dev'));
+gulp.task('build:web:dev', gulp.series('relay:copy', 'copy:build:web', 'webpack:build:web:dev'));
 
 gulp.task('serve:server', (callback) => {
   const app = require('./scripts/server-app');
