@@ -1,9 +1,8 @@
-require('babel-register')();
 require('ignore-styles').default(['.css']);
 
-var jsdom = require('jsdom').jsdom;
-
-var exposedProperties = ['window', 'navigator', 'document'];
+const path = require('path');
+const jsdom = require('jsdom').jsdom;
+const exposedProperties = ['window', 'navigator', 'document'];
 
 global.document = jsdom('');
 global.window = document.defaultView;
@@ -19,9 +18,22 @@ global.navigator = {
   platform: 'linux',
 };
 
-documentRef = document;
-
-var enzyme = require('enzyme');
-var Adapter = require('enzyme-adapter-react-16');
-
+const enzyme = require('enzyme');
+const Adapter = require('enzyme-adapter-react-16');
 enzyme.configure({ adapter: new Adapter() });
+
+// Override "require('config')" to return the test-config name.
+// Webpack won't do this, because when target=node it just outputs
+// as normal.
+//
+// TODO use a different test runner -- e.g., jest.mock() would be
+// cleaner.
+const Module = require('module');
+const originalResolve = Module._resolveFilename;
+Module._resolveFilename = (name, ...args) => {
+  if (name === 'config') {
+    return path.resolve(__dirname, 'config.js');
+  } else {
+    return originalResolve.call(this, name, ...args);
+  }
+}
